refactor: more refactoring of conftest
