+++ tests/data_context/test_configuration_storage_pre_v013.py
@@ -1,7 +1,11 @@
 import logging
 import os
+import shutil
 
 import pytest
+
+import great_expectations as ge
+from great_expectations.data_context.util import file_relative_path
 
 logger = logging.getLogger(__name__)
 
@@ -10,6 +14,51 @@ def read_config_file_from_disk(config_filepath):
     with open(config_filepath) as infile:
         config_file = infile.read()
         return config_file
+
+
+@pytest.fixture
+def data_context_parameterized_expectation_suite_no_checkpoint_store_with_usage_statistics_enabled(
+    tmp_path_factory,
+):
+    """
+    This data_context is *manually* created to have the config we want, vs
+    created with DataContext.create()
+    """
+    project_path = str(tmp_path_factory.mktemp("data_context"))
+    context_path = os.path.join(project_path, "great_expectations")
+    asset_config_path = os.path.join(context_path, "expectations")
+    fixture_dir = file_relative_path(__file__, "../test_fixtures")
+    os.makedirs(
+        os.path.join(asset_config_path, "my_dag_node"),
+        exist_ok=True,
+    )
+    shutil.copy(
+        os.path.join(
+            fixture_dir, "great_expectations_basic_with_usage_stats_enabled.yml"
+        ),
+        str(os.path.join(context_path, "great_expectations.yml")),
+    )
+    shutil.copy(
+        os.path.join(
+            fixture_dir,
+            "expectation_suites/parameterized_expectation_suite_fixture.json",
+        ),
+        os.path.join(asset_config_path, "my_dag_node", "default.json"),
+    )
+    os.makedirs(os.path.join(context_path, "plugins"), exist_ok=True)
+    shutil.copy(
+        os.path.join(fixture_dir, "custom_pandas_dataset.py"),
+        str(os.path.join(context_path, "plugins", "custom_pandas_dataset.py")),
+    )
+    shutil.copy(
+        os.path.join(fixture_dir, "custom_sqlalchemy_dataset.py"),
+        str(os.path.join(context_path, "plugins", "custom_sqlalchemy_dataset.py")),
+    )
+    shutil.copy(
+        os.path.join(fixture_dir, "custom_sparkdf_dataset.py"),
+        str(os.path.join(context_path, "plugins", "custom_sparkdf_dataset.py")),
+    )
+    return ge.data_context.DataContext(context_path)
 
 
 def test_preserve_comments_in_yml_after_adding_datasource(
